add new tests for class Calculator
