+++ src/test/java/com/sbrf/reboot/calculator/CalculatorTest.java
@@ -1,28 +1,35 @@
 package com.sbrf.reboot.calculator;
 
+import com.sbrf.reboot.Calculator;
 import org.junit.jupiter.api.Test;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.assertThrows;
 
 class CalculatorTest {
     @Test
     void getAddition() {
         assertEquals(9, Calculator.getAddition(4, 5));
+        assertThrows(ArithmeticException.class, () -> Calculator.getAddition(Long.MAX_VALUE, 1));
     }
 
     @Test
     void getSubtraction() {
         assertEquals(-1, Calculator.getSubtraction(4, 5));
+        assertThrows(ArithmeticException.class, () -> Calculator.getSubtraction(Long.MIN_VALUE, 1));
+
     }
 
     @Test
     void getMultiplication() {
         assertEquals(20, Calculator.getMultiplication(4, 5));
+        assertThrows(ArithmeticException.class, () -> Calculator.getMultiplication(Long.MAX_VALUE, 2));
     }
 
     @Test
     void getDivision() {
         assertEquals(3, Calculator.getDivision(9, 3));
+        assertThrows(ArithmeticException.class, () -> Calculator.getDivision(5, 0));
     }
 
     @Test
@@ -41,6 +48,7 @@ class CalculatorTest {
     void getAbs() {
         assertEquals(12, Calculator.getAbs(12));
         assertEquals(5, Calculator.getAbs(-5));
+        assertThrows(ArithmeticException.class, () -> Calculator.getAbs(Long.MIN_VALUE));
     }
 
     @Test
